+++ ydb/core/blobstorage/nodewarden/distconf_invoke_state_storage.cpp
@@ -2,7 +2,7 @@
 #include "distconf_selfheal.h"
 
 #include <ydb/core/base/statestorage.h>
-#include <ydb/core/base/statestorage_nodestate.h>
+#include <ydb/core/base/nodestate.h>
 
 
 namespace NKikimr::NStorage {
