+++ ydb/core/blobstorage/nodewarden/distconf_statestorage_config_generator.h
@@ -2,7 +2,7 @@
 
 #include "distconf.h"
 
-#include <ydb/core/base/statestorage_nodestate.h>
+#include <ydb/core/base/nodestate.h>
 
 namespace NKikimr::NStorage {
